moving the python files around, slightly cleaner landing page
