+++ python/polyglot.py
